+++ packages/pi-little-helpers/CHANGELOG.md
@@ -13,6 +13,12 @@ system4d:
 
 All notable changes to this project should be documented here.
 
+## Unreleased
+
+### Added
+
+- Add `sidequest` extension command to fork the current Pi session into a new Ghostty tab when supported, with fallback to a new Ghostty window.
+
 ## [0.2.0](https://github.com/tryingET/pi-little-helpers/compare/v0.1.3...v0.2.0) (2026-02-27)
 
 ### Changed